split code to more readable files
