+++ src/expressions/api/deserialize/type-store.ts
@@ -1,0 +1,4 @@
+import type { ExpressionNode, NodeExpressionWithType } from '../expression.js';
+
+
+export const expressionTypes: Map<string, NodeExpressionWithType<ExpressionNode>> = new Map();
