+++ src/expressions/index.ts
@@ -1,3 +1,4 @@
 export * from './scope/index.js';
 export * from './api/index.js';
 export * from './v8/index.js';
+export * from './visit/index.js';
